Use METADATA, not metadata.json
