+++ python/requirements.txt
@@ -1,3 +1,4 @@
+packaging==20.0
 pip==9.0.3
 setuptools==38.2.4
 wheel==0.30.0a0
